triggersignals/requirements.txt
telethon
python-dotenv
schedule